feat(souffle): Rewrite the library to support mappings to AST positions
