+++ src/internals/souffle/index.ts
@@ -1,0 +1,10 @@
+export { Fact, FactValue, FactType } from "./fact";
+export {
+  RelationName,
+  RelationArgName,
+  RelationIO,
+  Relation,
+} from "./relation";
+export { Atom, RuleHead, RuleName, RuleBody, Rule } from "./rule";
+export { Context } from "./context";
+export { Executor } from "./executor";
